moved gui to qtgui, added hypothetical CEGUI support, added QML for main window and preview frames. builds, runs, not functional
