+++ qtgui/UiApplication.cpp
