+++ qtgui/UiApplication.h
