+++ qtgui/resources/qtquickcontrols2.conf
